backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
twilio==8.13.0
google-generativeai==0.3.2
requests==2.31.0
python-multipart==0.0.9
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

